requirements.txt
runpod==1.6.0
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.200
transformers==4.35.0
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
